type change

+++ backend/telekom-analiz-sistemi/telekom/src/main/java/com/pia/telekom/entity/CollectionAction.java
@@ -21,7 +21,7 @@ public class CollectionAction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "action_id")
-    private Long actionId;
+    private int actionId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "invoice_id", nullable = false)
